feat: skip existing files (#1)

+++ src/utils/mod.rs
@@ -1,5 +1,5 @@
 use colored::Colorize;
-use std::fs::File;
+use std::{fs::File, path::Path};
 use error_chain::error_chain;
 
 error_chain! {
@@ -11,6 +11,10 @@ error_chain! {
 
 pub async fn download(url: String, path: String) -> Result<()> {
     let file = format!("{}/{}.png", path, url.split_once("image/").unwrap().1.split_once("?").unwrap().0);
+    if Path::new(&file).exists() {
+        println!("Skipping {} because it already downloaded", file.bright_blue());
+        return Ok(());
+    }
     println!("Downloading from {} to {}", url.bright_blue(), file.green()); 
     let resp = reqwest::get(url).await.expect("request failed");
     let mut out = File::create(file).expect("failed to create file");
